unify base info screen for add estate and filter estates
