+++ lib/screens/add_estate/introduction.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:saja/resources/colors.dart';
 import 'package:saja/resources/strings.dart';
-import 'package:saja/screens/add_estate/estate_delegation_type.dart';
+import 'package:saja/screens/base_info.dart';
 import 'package:saja/services/navigation/app_navigator.dart';
 import 'package:saja/widgets/custom_button.dart';
 
@@ -52,7 +52,8 @@ class AddEstateIntroductionScreen extends StatelessWidget {
             icon: Icons.keyboard_arrow_left_sharp,
             iconPadding: 5,
             onPressed: () {
-              AppNavigator.pushScreen(context, EstateDelegationTypeScreen());
+              AppNavigator.pushScreen(context,
+                  BaseInfoScreen(nextScreenType: NextScreenType.AddEstate));
             },
           )
         ],
